fix(review): resolve decisions #1, #2, #3 from PR #20 review
- #1 Merge routes/listing.ts and routes/file.ts into routes/api.ts
     (single module owns the entire /api/${prefix}/* tree, sendListing
     becomes a private helper). Eliminates the cross-module export
     forced by the Fastify duplicate-route constraint.
- #2 Replace FileEntry interface with a discriminated union
     ({kind: "file", size: number} | {kind: "dir"}). Drops the
     `size: 0` sentinel; consumers now narrow on kind before reading
     size. mcp.ts and listing-page already branch on kind, so behavior
     is unchanged.
- #3 Add stat-and-redirect for bare-directory URLs in the wildcard
     viewer branch. /<prefix>/<sub> (no trailing slash) where <sub>
     resolves to a directory now 302s to /<prefix>/<sub>/, landing
     users on the listing instead of a 404'ing viewer shell. The stat
     is jailed by resolveSafePath; on traversal/missing we silently
     fall through to preserve existing error semantics.

Verified: bare-dir 302, clean-file 200, .md redirect, traversal 403,
SSE for nested files, slash-form listings — all pass against the
fixture matrix.
